Re-structure the test suite
