+++ test/core/preconditioning.jl
