+++ test/core/priors.jl
